refactor(runtime): extract snapshot store module from runtime monolith (#3090)
